# apps/server/src/docs/openapi-stub.yaml
openapi: 3.0.3
info:
  title: Minddeck API
  version: 0.1.0
  description: >
    Minimal but truthful spec for the current backend. Uses JWT bearer auth.
    Lists return arrays; pagination is exposed via response headers.
servers:
  - url: http://localhost:5000
...
paths:
...
  /decks:
    get:
      summary: List my decks
      tags: [decks]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Array of decks
          headers:
            X-Total-Count:
              { description: Total rows, schema: { type: integer } }
            X-Page: { description: Current page, schema: { type: integer } }
            X-Limit: { description: Page size, schema: { type: integer } }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Deck" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Create deck
      tags: [decks]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeckInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deck" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /cards:
    post:
      summary: Create card
      tags: [cards]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCardInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Card" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
      example: { error: "Invalid deck id" }
...
    Deck:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        user: { type: string, description: "Owner user id" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateDeckInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
    Card:
      type: object
      properties:
        _id: { type: string }
        front: { type: string }
        back: { type: string }
        deck: { type: string, description: "Deck id" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateCardInput:
      type: object
      required: [front, back, deckId]
      properties:
        front: { type: string, minLength: 1 }
        back: { type: string, minLength: 1 }
        deckId: { type: string }
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            invalid_id: { value: { error: "Invalid deck id" } }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            missing: { value: { error: "Missing or invalid token" } }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            not_owner: { value: { error: "Not your card" } }
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            dup_title: { value: { error: "Deck title already exists" } }
...
security:
  - bearerAuth: []